+++ src/Universum/Functor/Reexport.hs
@@ -4,10 +4,12 @@ module Universum.Functor.Reexport
        ( module Control.Arrow
        , module Data.Bifunctor
        , module Data.Functor
+       , module Data.Functor.Compose
        , module Data.Functor.Identity
        ) where
 
 import Control.Arrow ((&&&))
 import Data.Bifunctor (Bifunctor (..))
 import Data.Functor (Functor (..), void, ($>), (<$>))
+import Data.Functor.Compose (Compose (..))
 import Data.Functor.Identity (Identity (..))