+++ CHANGES.md
@@ -3,6 +3,8 @@
 
 * [#53](https://github.com/serokell/universum/issues/53):
   Add `doctest` to `universum`. Also imporove and fix documentation.
+* [#117](https://github.com/serokell/universum/issues/117):
+  Drop the support of `GHC-8.0.1`.
 
 1.0.3
 =====